Move all state to Redux
In normal, idiomatic Redux, it's fine to let things like text input
values, or UI toggles, use React component state instead of global
Redux state.

The default TodoMVC stores the text value for the current Todo,
as well as the filter value for visibilityFilter, in component state.

However, when using ReduxVCR, we only persist actions that make it
to the store.

This _optional_ commit allows us to record and replay the stuff that
users write in the "todo" text input before hitting submit. Without
this commit, ReduxVCR would still work perfectly, but todos would
magically appear instead of being typed in, and we wouldn't capture
any toggles to the visibility filter.

It is up to you, when using ReduxVCR, how granular you want to be
with your recording :)